examples/deployment/requirements.txt
google-adk>=0.1.0
uvicorn
fastapi
# markov-agent>=0.1.0  # Uncomment if available on PyPI or install from source
